requirements.txt
astropy==5.1.1
cffi==1.15.1
cffi>=1.0.0
colorlog==6.7.0
corner==2.2.1
emcee==3.1.3
joblib==1.2.0
jsonschema==4.16.0
matplotlib==3.6.1
numpy==1.23.4

# optional for tools combine
# git+https://github.com/cylammarco/SpectRes

# Development Requirements
pre-commit==2.20.0
pytest==7.2.0
pytest-cov==4.0.0
python-dateutil==2.8.2
scikit-image==0.19.3
scipy==1.10.0
sphinx==5.3.0
tqdm==4.64.1
wget==3.2
joblib==0.13.2
pytest==5.0.1
jsonschema==3.0.1
Pillow==9.3.0